BSHS-2: update-hashsum-file
A script for updating a hashfile produced by sha256sum and similar.
It works on a directory. First, we find any files older than the
existing hashfile. Then, we compute new hashes, and either append them,
or replace existing hashes.

Note that for now the script does not account for any files deleted
from the folder, but with hashes still in the hashfile.

Implements https://github.com/valera-rozuvan/bash-scripts/issues/2